Api cleanup

+++ packages/api/src/services/social/friend.service.ts
@@ -3,26 +3,25 @@ import { err, ok, Result } from "neverthrow";
 
 import { CloudFront } from "@oppfy/cloudfront";
 import type { Database } from "@oppfy/db";
+import { FollowStatus } from "@oppfy/db/utils/query-helpers";
+import { SQS } from "@oppfy/sqs"; // Added SQS import
 
 import * as FriendErrors from "../../errors/social/friend.error";
 import { ProfileNotFound } from "../../errors/user/profile.error";
-import {
-  BidirectionalUserIdsparams,
-  DirectionalUserIdsParams,
-  FollowStatus,
-  PaginatedResponse,
-  PaginationParams,
-} from "../../interfaces/types";
-import { HydratedProfile } from "../../models";
+import { Profile } from "../../models";
 import { FollowRepository } from "../../repositories/social/follow.repository";
 import { FriendRepository } from "../../repositories/social/friend.repository";
 import { ProfileRepository } from "../../repositories/user/profile.repository";
 import { UserRepository } from "../../repositories/user/user.repository";
 import { TYPES } from "../../symbols";
-
-type SocialProfile = HydratedProfile & {
-  followStatus: FollowStatus;
-};
+import {
+  BidirectionalUserIdsparams,
+  DirectionalUserIdsParams,
+  PaginatedResponse,
+  PaginationParams,
+  SelfOtherUserIdsParams,
+} from "../../types";
+import { Hydrate, hydrateProfile } from "../../utils";
 
 interface PaginateByUserIdWithSelfUserIdParams extends PaginationParams {
   selfUserId: string;
@@ -48,6 +47,7 @@ export class FriendService {
     private readonly userRepository: UserRepository,
     @inject(TYPES.ProfileRepository)
     private readonly profileRepository: ProfileRepository,
+    @inject(TYPES.SQS) private readonly sqs: SQS, // Injected SQS service
   ) {}
 
   /**
@@ -70,11 +70,19 @@ export class FriendService {
       return err(new FriendErrors.CannotFriendSelf(senderUserId));
     }
 
-    const recipientProfile = await this.profileRepository.getProfile({
-      userId: recipientUserId,
-    });
+    const [recipientProfile, senderProfile] = await Promise.all([
+      this.profileRepository.getProfile({
+        userId: recipientUserId,
+      }),
+      this.profileRepository.getProfile({
+        userId: senderUserId,
+      }),
+    ]);
+
     if (recipientProfile === undefined)
       return err(new ProfileNotFound(recipientUserId));
+    if (senderProfile === undefined)
+      return err(new ProfileNotFound(senderUserId)); // Should not happen if called by authenticated user
 
     return await this.db.transaction(async (tx) => {
       const [
@@ -162,6 +170,13 @@ export class FriendService {
           );
         }
 
+        // Friend request accepted implicitly by sending a request back
+        await this.sqs.sendFriendAcceptedNotification({
+          senderId: senderUserId,
+          recipientId: recipientUserId,
+          username: senderProfile.username,
+        });
+
         return ok();
       }
 
@@ -170,13 +185,25 @@ export class FriendService {
         tx,
       );
 
+      // Send notification for new friend request
+      await this.sqs.sendFriendRequestNotification({
+        senderId: senderUserId,
+        recipientId: recipientUserId,
+        username: senderProfile.username,
+      });
+
       if (!isFollowing && !isFollowRequested) {
-        await this.followRepository[
+        const action =
           recipientProfile.privacy === "public"
             ? "createFollower"
-            : "createFollowRequest"
-        ]({ senderUserId, recipientUserId }, tx);
+            : "createFollowRequest";
+
+        await this.followRepository[action](
+          { senderUserId, recipientUserId },
+          tx,
+        );
       }
+
       if (isRecipientFollowRequested) {
         await this.followRepository.deleteFollowRequest(
           { senderUserId: recipientUserId, recipientUserId: senderUserId },
@@ -218,75 +245,89 @@ export class FriendService {
    * Accepts an incoming friend request, establishing a friendship and ensuring mutual follows.
    */
   async acceptFriendRequest({
-    senderUserId,
-    recipientUserId,
-  }: DirectionalUserIdsParams): Promise<
-    Result<void, FriendErrors.RequestNotFound>
+    selfUserId,
+    otherUserId,
+  }: SelfOtherUserIdsParams): Promise<
+    Result<void, FriendErrors.RequestNotFound | ProfileNotFound>
   > {
+    const selfProfile = await this.profileRepository.getProfile({
+      userId: selfUserId,
+    });
+    if (selfProfile === undefined) return err(new ProfileNotFound(selfUserId)); // Should not happen
+
     return await this.db.transaction(async (tx) => {
+      // Check that a friend request exists from `otherUserId` -> `selfUserId`
       const isRequested = await this.friendRepository.getFriendRequest(
-        { senderUserId, recipientUserId },
+        { senderUserId: otherUserId, recipientUserId: selfUserId },
         tx,
       );
       if (!isRequested) {
-        return err(
-          new FriendErrors.RequestNotFound(recipientUserId, senderUserId),
-        );
+        return err(new FriendErrors.RequestNotFound(selfUserId, otherUserId));
       }
 
+      // Delete that friend request and create the friend record
       await this.friendRepository.deleteFriendRequest(
-        { senderUserId, recipientUserId },
+        { senderUserId: otherUserId, recipientUserId: selfUserId },
         tx,
       );
       await this.friendRepository.createFriend(
-        { userIdA: senderUserId, userIdB: recipientUserId },
+        { userIdA: otherUserId, userIdB: selfUserId },
         tx,
       );
 
+      // Send notification for accepted friend request
+      await this.sqs.sendFriendAcceptedNotification({
+        senderId: selfUserId,
+        recipientId: otherUserId,
+        username: selfProfile.username,
+      });
+
+      // Check if there are any follow requests both ways and remove them
       const [isFollowRequested, isRecipientFollowRequested] = await Promise.all(
         [
           this.followRepository.getFollowRequest(
-            { senderUserId, recipientUserId },
+            { senderUserId: otherUserId, recipientUserId: selfUserId },
             tx,
           ),
           this.followRepository.getFollowRequest(
-            { senderUserId, recipientUserId },
+            { senderUserId: selfUserId, recipientUserId: otherUserId },
             tx,
           ),
         ],
       );
       if (isFollowRequested) {
         await this.followRepository.deleteFollowRequest(
-          { senderUserId, recipientUserId },
+          { senderUserId: otherUserId, recipientUserId: selfUserId },
           tx,
         );
       }
       if (isRecipientFollowRequested) {
         await this.followRepository.deleteFollowRequest(
-          { senderUserId, recipientUserId },
+          { senderUserId: selfUserId, recipientUserId: otherUserId },
           tx,
         );
       }
 
+      // Ensure both are now following each other
       const [isFollowing, isRecipientFollowing] = await Promise.all([
         this.followRepository.getFollower(
-          { senderUserId, recipientUserId },
+          { senderUserId: otherUserId, recipientUserId: selfUserId },
           tx,
         ),
         this.followRepository.getFollower(
-          { senderUserId, recipientUserId },
+          { senderUserId: selfUserId, recipientUserId: otherUserId },
           tx,
         ),
       ]);
       if (!isFollowing) {
         await this.followRepository.createFollower(
-          { senderUserId, recipientUserId },
+          { senderUserId: otherUserId, recipientUserId: selfUserId },
           tx,
         );
       }
       if (!isRecipientFollowing) {
         await this.followRepository.createFollower(
-          { senderUserId, recipientUserId },
+          { senderUserId: selfUserId, recipientUserId: otherUserId },
           tx,
         );
       }
@@ -362,7 +403,14 @@ export class FriendService {
     pageSize = 10,
     selfUserId,
   }: PaginateByUserIdWithSelfUserIdParams): Promise<
-    Result<PaginatedResponse<SocialProfile>, never>
+    Result<
+      PaginatedResponse<
+        Hydrate<Profile<"onboarded">> & {
+          followStatus: FollowStatus;
+        }
+      >,
+      never
+    >
   > {
     const rawProfiles = await this.friendRepository.paginateFriends({
       userId,
@@ -372,7 +420,7 @@ export class FriendService {
     });
 
     const hydratedProfiles = rawProfiles.map((profile) => ({
-      ...this.cloudfront.hydrateProfile(profile),
+      ...hydrateProfile(profile),
       followStatus: profile.followStatus,
     }));
 
@@ -397,7 +445,7 @@ export class FriendService {
     cursor,
     pageSize = 10,
   }: PaginateByUserIdParams): Promise<
-    Result<PaginatedResponse<HydratedProfile>, never>
+    Result<PaginatedResponse<Hydrate<Profile<"onboarded">>>, never>
   > {
     const rawProfiles = await this.friendRepository.paginateFriendRequests({
       userId,
@@ -406,7 +454,7 @@ export class FriendService {
     });
 
     const hydratedProfiles = rawProfiles.map((profile) =>
-      this.cloudfront.hydrateProfile(profile),
+      hydrateProfile(profile),
     );
 
     const hasMore = rawProfiles.length > pageSize;
